tests/wpt/tests/fenced-frame/battery_status.https.html

<!DOCTYPE html>
<title>Battery status API test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
promise_test(async () => {
  const uuid = token();
  const frame = attachFencedFrame(
    generateURL(`resources/get_battery.html`, [uuid]));
  const result = await nextValueFromServer(uuid);
  assert_equals(result, "NotAllowedError",
                "battery status API should cause exception");
}, 'battery status should not be read in the fenced frame.');
</script>
</body>
